In Multiboot use compiledate as image date and imagedevbuild as is
In most images including recently OpenPLi imagedevbuild in the boxinfo
included the date from the image. But in Vix it was a release number.
Yesterday OpenPLi adapted to the OpenVix format. By using compiledate we
always get a date correctly. In images where they have the date here it
is now shown as version number.